+++ .gitignore
@@ -1,3 +1,1 @@
 __pycache__
-ui_tests/features/steps/helpers/__pycache__
-api_tests/features/reports
