+++ tests/features/environment.py
@@ -50,7 +50,7 @@ def after_all(context):
     time.sleep(5)
     os.system('./../tools/cm selenoid stop')
     os.system('./../tools/cm selenoid-ui stop')
-    os.system('cp ~/.aerokube/selenoid/video/*.mp4 ~/Vídeos/')
+    os.system('cp ~/.aerokube/selenoid/video/*.mp4 ~/Videos/')
 
     print(" ")
     print("--------------------------------------------")
